Low: daemons: Support cluster-wide sync points for multi IPC messages.
Supporting cluster-wide sync points means attrd_expect_confirmations
needs to be called, and then attrd_send_message needs "true" as a third
argument.  This indicates attrd wants confirmations back from all its
peers when they have applied the update.

We're already doing this at the end of attrd_client_update for
single-update IPC messages, and handling it for multi-update messages is
a simple matter of breaking that code out into a function and making
sure it's called.

Note that this leaves two other spots where sync points still need to be
dealt with:

* An update message that uses a regex.  See
  https://projects.clusterlabs.org/T600 for details.

* A multi-update IPC message in a cluster where that is not supported.
  See https://projects.clusterlabs.org/T601 for details.