Feature: libcrmcluster: New crm_join_nack_quiet crm_join_phase enum val
This phase indicates that we should not send a nack message to the
client; we carry out the rest of the nack procedure normally. Clients
with a feature set older than 3.17.0 (coming in another commit) may
respawn infinitely upon receiving a nack message.

In practice, that should only affect clients that are even older, since
verify_feature_set() runs on the client side and causes the node to
shut down with respawn inhibited. For safety, we'll draw the line at
3.17.0 anyway.

It would be much more convenient to add a feature_set member to
crm_node_t. However, since the struct is public API, the best we can do
is add an crm_join_phase enum value.

Ref T453

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>