API: libpe_status: add pe_rsc_detect_loop constant
The scheduler's resource assignment code uses pe_rsc_allocating to detect loops
in the assignment. This constant was overloaded elsewhere to detect loops in
other situations, but that is undesirable because it prevents pe_rsc_allocating
from being used to detect whether a resource is in the process of being
assigned.

Add pe_rsc_detect_loop for the other situations, and use it there.