Feature: python: Add the very beginnings of a pacemaker python library.
The purpose of this library is currently to be a place for consolidating
common code among our regression tests.  It may eventually grow to
include an interface to pacemaker itself, but that's not yet planned.

The presence of the setup.py.in file is required to make sure that the
library can be built as part of package builds.  It will also be
necessary should we ever want to upload this library to pip.