Low: tools: Remove crm_log_args() calls from cibadmin
Currently the crm_log_args() calls don't even work. It logs at notice
level, but the cibadmin log level is always crit at that point. The
verbose flags have not yet bumped the log level.

It's not hard to fix that by logging args (if appropriate) after bumping
the log level, but it clutters the code for no apparent benefit. The
only place it will ever be logged is to stderr when manually running
cibadmin with four or more -V flags, or possibly by resource agents.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>