Low: controller: Avoid error if a join request is received after fencing
If a node is fenced while it's joining the cluster, it's possible for
the DC to receive its join request after noticing it's down, leading to
error logs about the node not being a member and the request being
rejected. We shouldn't log an error in this scenario since we know the
node went down recently.

This scenario has not been reproducible despite a great deal of trying.
Nonetheless, this change only affects log severity, so it's reasonable
to implement it in case the situation reappears.

Closes T456

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>