Feature: libpe_status: New cluster-id XML node element attribute
The new cluster-id attribute of a <node> element sets the node's
cluster-layer ID. Currently, this is the same role that the id attribute
fills. In the future, the plan is to constrain the id attribute to valid
XML IDs.

This commit also causes both the uname attribute and the cluster-id
attribute to default to the id attribute's value if unset.

Ref T290

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>