Fix: scheduler: prevent a leftover pending monitor from causing unexpected stop of other instances
As far as I can tell, the test `leftover-pending-monitor` exposes an
issue when deciding whether to add a `pe_order_runnable_left` flag for
the ordering related to a probe, where it's supposed to check whether
there's any parent resource with `running_on` rather than the instance
itself.

According to 7d37dc58d upon unpacking, with the leftover pending monitor
of stateful-1 on node-1, stateful-1:0 is exclusively assigned to node-1
and determined being stopped there. With the previous logic, given that
stateful-1:0 has no `running_on`, despite the fact that another instance
is already running on node-3, `pe_order_runnable_left` flag is enforced
for the ordering between `stateful-1:0_monitor_0` and
`promotable-1_start_0`. But then stateful-1:0_monitor_0 is marked
unrunnable because node-2 is pending, and then so is
promotable-1_start_0...