Doc: libcrmcommon: Remove incorrect comment in patchset.c
The example diff XML in patchset.c was added 9 years ago in f5be869c and
has been incorrect for a long time. For example, format is an integer,
not a float, and the correct attribute is path, not xpath.

Cleaning this up is an option. However, since we've just added a schema
for diff, this now seems redundant.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>