Fix: controller: Don't nack joining node due to old CIB
In the finalization stage, we sync the max_generation CIB across the
cluster. It's possible for the sync_from() operation to return a
-pcmk_err_old_data error ("Update was older than existing
configuration"). That error is fairly normal (hence treating it as a
warning already), and it shouldn't cause us to call a new election and
nack the max_generation node.

Regression in ae899e8.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>