Low: libcrmcommon: Don't use aliases in XML logging functions
The aliases aren't very informative in most or all of the current
callers. For example, some callers pass an aliased function name while
using the true file name and line number, which is confusing and an
abuse of the intent of do_crm_log_alias(). Other callers pass the name
of a function like cib_perform_op(), which is widely used and generic in
its functionality so that it's not much more informative than the true
function name. Even if there are a more legitimate uses, we're not
losing much here.

Alias logs also complicate the implementation of an upcoming formatted
output feature for patchsets.

The affected public functions are as follows:
* xml_log_patchset()
* xml_log_changes()
* log_data_element()

The signatures can't change, but the file, function, and line
arguments are now ignored.

There are a few other affected static and internal functions.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>