Refactor: libcrmcommon: Reduce pcmk__cli_help.
This function cannot be fully removed until all tools are converted to
using formatted output.  However, the only times it's used anymore are
for 'v' and '!' (other uses happen in other functions that are about to
be removed).  So, just strip it down to doing those things.

We need to remove the rest of this code in order to get rid of
references to several variables that will only be set by code that's
about to be removed.