Low: libcrmcommon: Warn if deprecated command line formats are used.
We want to get rid of pcmk__cmdline_preproc eventually, but before we
can do that we need to get rid of its reason for existing.  Currently,
that is only to support the fact that getopt supports constructions like
"-x100" while glib does not.  glib expects "-x 100", so this function
converts the command line arguments into a format it can handle.

Print out a warning if the deprecated format is used so we can take care
of this with a future release.

Ref T92