Refactor: scheduler: clarify arguments when checking parameters
Drop a redundant argument to the param check callback, and make a couple of the
callback arguments const.

Also, clarify pointer arguments in the call chain for checking parameters.
Mark as [in,out] or const where appropriate. This is not comprehensive but
focuses on functions with doxygen blocks and some functions those call.