Low: scheduler: order cloned fence device probes same as other clones
We order probes before start or reload -- of the probed resource in most cases,
but of the entire clone if the probed resource is a clone instance.

Previously, there was an exception for instances of cloned fence devices when
the cluster uses unfencing. This exception doesn't make much sense, especially
considering that we don't even have good support for cloned fence devices, so
just drop it.

The exception was added by faf44d811 and was likely intended for non-cloned
fence devices, which is redundant with the non-clone test.

No regression tests change.