Refactor: libcrmcommon: Remove internal uses of log_data_element()
Replace each call with the code that log_data_element() would have run
for that caller. For xml_log_changes(), the replacements are trivial,
because log_data_element() would only have run a couple of meaningful
lines of code.

For xml_log_patchset_v1(), the replacement is somewhat larger, but this
is the only caller of the recursive portion of log_data_element -- so it
needs to be done in only one place.

The replacement in do_crm_log_xml() will come in the next commit. It
requires a slight API change: the macro is reimplemented as a function.

This is the final step before deprecating log_data_element().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>