Refactor: cts,python: Add a base Test class.
For the moment, all it's got is __init__, exitcode, and logpath.  Lots
more will go into this class in the future as I combine cts-exec and
cts-fencing into a single library.

I've also renamed a bunch of attributes to be private, which seems more
fitting for how they are used.