Revert "Fix: libpacemaker: ensure any pending recurring monitor gets updated if it fails"
This reverts commit 11689380bb663bde3699093248039aa9479082f5.

If a recurring action fails its first run, the CIB resource history will
contain a recurring action entry with a pending status and a last_failure.
The resource will wrongly be displayed as "Monitoring".

The reverted commit tried to avoid that by always overwriting the recurring
action entry with failures. However the scheduler relies on recurring failures
being only in last_failure to properly calculate the resource state. This is
particularly important for remote connection resources with a reconnect
interval, as the remote node will be wrongly fenced if a recurring monitor
entry contains an expired failure.

So, revert the original commit. The display issue will be addressed in a
different manner later.