Fix: scheduler: ignore node when getting resource inside container
This looks like a logic error from the beginning. The container instance passed
to pcmk__get_rsc_in_container() has already been determined to be "compatible,"
meaning it is on the proper node and has the proper role for interleaving.

replica->node is the guest node created by the replica's remote connection,
and therefore will never match the node where container actions are being
interleaved.

No regression tests change, which indicates we are never interleaving the inner
resource's actions in the current tests.