Feature: tools: Add pattern support to attrd_updater.
This adds support for patterns for all commands except queries, which
are more difficult due to libxml (see comment in command_query in
tools/crm_attribute.c).

Adding this also helps to verify that sync points work with regexes, as
well as brings it a little closer to crm_attribute so the two tools can
one day be merged.