Refactor: various: Use pcmk__output_t in pcmk__xml_log_changes()
And rename it to pcmk__xml_show_changes().

pcmk__xml_show_changes() had already been updated to create a
pcmk__output_t object that eventually gets passed to pcmk__xml_show().
This commit allows us to reuse an existing pcmk__output_t object
(currently only done in based_callbacks.c:process_ping_reply()).

An alternative would be to allow pcmk__xml_show_changes() to accept a
pcmk__output_t argument and a log_level argument: if the output argument
is NULL, create a new temporary one based on the level argument. That
would simplify most of the callers at the expense of muddying the
interface and implementation of pcmk__xml_show_changes().

Note that there's no current need to use pcmk__xml_show_changes() with
any output format other than text and log, so we just use out->info().
Formatting the output for XML would be awkward, so I'd rather cross that
bridge if and when we come to it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>