Refactor: cts: Clarify usage of self._stonith_* in CTS tests.
In cts-fencing.in, self._stonith_* is really self._daemon_* since that's
the thing being tested.  We can then use the same variable names in
cts-exec.in for the thing being tested there, making the two programs
more similar and easier to combine.

In cts-exec.in, we still additionally need to keep track of a stonith
process.  So continue to use self._stonith_process there for that
purpose.