Fix: scheduler: count only containers' active nodes for bundles
active_node() has a couple of problems when used with bundles.

It counts all nodes where any of the bundle's implicit resources are active,
including the guest node created by the implicit remote connection (if any).
Even if only a single replica is active, crm_resource will refuse to move or
ban it because the active node count is higher than 1.

Second, it checks for the partial migration source of the bundle itself, which
will never exist. If the bundle has only a single active replica, we should
check for the partial migration source of that replica.

Fixes T619