Refactor: libcrmcommon: New pcmk__output_select_rc() function
Per the doxygen, this is a convenience function to simplify deciding how
to update a saved return code based on the latest operation using a
pcmk__output_t object. Basically, it tries to keep the most meaningful
return code from a series of operations. This will be used in upcoming
commits.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>