High: libcrmcommon: Fix handling node=NULL in pcmk__attrd_api_query.
According to the header file, if node is NULL, pcmk__attrd_api_query
should query the value of the given attribute on all cluster nodes.
This is also what the server expects and how attrd_updater is supposed
to work.

However, pcmk__attrd_api_query has no way of letting callers decide
whether they want to query all nodes or whether they want to use the
local node.  We were passing NULL for the node name, which it took to
mean it should look up the local node name.  This calls
pcmk__node_attr_target, which probes the local cluster name and returns
that to pcmk__attrd_api_query.  If it returns non-NULL, that value will
then be put into the XML IPC call which means the server will only
return the value for that node.

In testing this was usually fine.  However, in pratice, the methods
pcmk__node_attr_target uses to figure out the local cluster node name
involves checking the OCF_RESKEY_CRM_meta_on_node environment variable
among others.

This variable was never set in testing, but can be set in the real
world.  This leads to circumstances where the user did "attrd_updater -QA"
expecting to get the values on all nodes, but instead only got the value
on the local cluster node.

In pacemaker-2.1.4 and prior, pcmk__node_attr_target was simply never
called if the node was NULL but was called otherwise.

The fix is to modify pcmk__attrd_api_query to take an option for
querying all nodes.  If that's present, we'll query all nodes.  If it's
not present, we'll look at the given node name - NULL means look it up,
anything else means just that node.

Regression in 2.1.5 introduced by eb20a65577