Feature: python: Allow overriding the environment in cts tests.
Functions like add_cmd and add_cmd_check_stdout now take an env
parameter that, if not None, contains things that will be added to the
environment before executing the test.  The env parameter should be a
dict.  It can contain things that already exist in the environment, in
which case they will override previous settings.