Fix: scheduler: Do not fence a pending node that doesn't have an uname in node state yet
If a joining peer makes the cluster acquire the quorum from corosync
meanwhile it has not joined CPG membership of pacemaker-controld yet,
it's possible that the created node_state entry doesn't have an uname
yet. Previously in that case, the node would be considered `UNCLEAN
(offline)` and get unnecessarily fenced before it got a chance to join
CPG yet.

The fix resolves that by recognizing the node as `pending` and waiting
for it to join CPG.