Refactor: tools: Scaffolding for formatted output in crm_shadow
Reviewing the changes for formatted output in crm_shadow will be
somewhat tedious. I'm breaking it into steps to make review easier.

This commit creates an output object that currently does nothing. It
registers formats and an empty array of message functions.

The letters in the pcmk__cmdline_preproc() call get rearranged for
better readability, which should have been done in the past.

Also make the pcmk__output_t object function-scope to main(). Moving to
libpacemaker will be easier if our helper functions won't rely on a
global output object.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>