Fix: scheduler: expired results shouldn't affect state
The main purpose of remap_operation() is to modify a result's execution status
and exit status. However it can also have side effects like changing a
resource's role or blocking a resource for certain failures. These side effects
should not occur if the result is an expired failure.

No regression tests change.